requirements.txt
paho-mqtt
watchdog==0.9.0
requests==2.22.0
